+++ packages/amazonq/src/extension.ts
@@ -4,7 +4,12 @@
  */
 
 import { CredentialsStore, LoginManager, authUtils, initializeAuth } from 'aws-core-vscode/auth'
-import { activate as activateCodeWhisperer, shutdown as shutdownCodeWhisperer } from 'aws-core-vscode/codewhisperer'
+import { Auth, AuthUtils, CredentialsStore, LoginManager, initializeAuth } from 'aws-core-vscode/auth'
+import {
+    activate as activateCodeWhisperer,
+    AuthUtil,
+    shutdown as shutdownCodeWhisperer,
+} from 'aws-core-vscode/codewhisperer'
 import { makeEndpointsProvider, registerGenericCommands } from 'aws-core-vscode'
 import { CommonAuthWebview } from 'aws-core-vscode/login'
 import {
@@ -34,6 +39,7 @@ import {
     Experiments,
     isSageMaker,
     Commands,
+    isAmazonInternalOs,
 } from 'aws-core-vscode/shared'
 import { ExtStartUpSources } from 'aws-core-vscode/telemetry'
 import { VSCODE_EXTENSION_ID } from 'aws-core-vscode/utils'
@@ -45,6 +51,7 @@ import { focusAmazonQPanel } from 'aws-core-vscode/codewhispererChat'
 import { activate as activateAmazonqLsp } from './lsp/activation'
 import { activate as activateInlineCompletion } from './app/inline/activation'
 import { isAmazonInternalOs } from 'aws-core-vscode/shared'
+import { hasGlibcPatch } from './lsp/client'
 
 export const amazonQContextPrefix = 'amazonq'
 
@@ -120,11 +127,17 @@ export async function activateAmazonQCommon(context: vscode.ExtensionContext, is
         extensionContext: context,
     }
 
-    await activateAmazonqLsp(context)
-
     // This contains every lsp agnostic things (auth, security scan, code scan)
     await activateCodeWhisperer(extContext as ExtContext)
 
+    if (
+        (Experiments.instance.get('amazonqLSP', true) || AuthUtil.instance.isInternalAmazonUser()) &&
+        (!isAmazonInternalOs() || (await hasGlibcPatch()))
+    ) {
+        // start the Amazon Q LSP for internal users first
+        // for AL2, start LSP if glibc patch is found
+        await activateAmazonqLsp(context)
+    }
     if (!Experiments.instance.get('amazonqLSPInline', false)) {
         await activateInlineCompletion()
     }