+++ packages/core/src/codewhisperer/activation.ts
@@ -95,6 +95,7 @@ import { SecurityIssueTreeViewProvider } from './service/securityIssueTreeViewPr
 import { setContext } from '../shared/vscode/setContext'
 import { syncSecurityIssueWebview } from './views/securityIssue/securityIssueWebview'
 import { detectCommentAboveLine } from '../shared/utilities/commentUtils'
+import { notifySelectDeveloperProfile } from './region/utils'
 
 let localize: nls.LocalizeFunc
 
@@ -374,6 +375,10 @@ export async function activate(context: ExtContext): Promise<void> {
                 if (auth.isIdcConnection()) {
                     await auth.notifySessionConfiguration()
                 }
+            }
+
+            if (auth.requireProfileSelection()) {
+                await notifySelectDeveloperProfile()
             }
         },
         { emit: false, functionId: { name: 'activateCwCore' } }