+++ packages/core/src/codewhisperer/service/transformByQ/transformFileHandler.ts
@@ -9,14 +9,14 @@ import * as os from 'os'
 import xml2js = require('xml2js')
 import * as CodeWhispererConstants from '../../models/constants'
 import { existsSync, readFileSync, writeFileSync } from 'fs' // eslint-disable-line no-restricted-imports
-import { BuildSystem, DB, FolderInfo, transformByQState } from '../../models/model'
+import { BuildSystem, DB, FolderInfo, TransformationType, transformByQState } from '../../models/model'
 import { IManifestFile } from '../../../amazonqFeatureDev/models'
 import fs from '../../../shared/fs/fs'
 import globals from '../../../shared/extensionGlobals'
 import { ChatSessionManager } from '../../../amazonqGumby/chat/storages/chatSession'
 import { AbsolutePathDetectedError } from '../../../amazonqGumby/errors'
 import { getLogger } from '../../../shared/logger/logger'
-import { isWin } from '../../../shared/vscode/env'
+import AdmZip from 'adm-zip'
 
 export function getDependenciesFolderInfo(): FolderInfo {
     const dependencyFolderName = `${CodeWhispererConstants.dependencyFolderName}${globals.clock.Date.now()}`
@@ -27,10 +27,53 @@ export function getDependenciesFolderInfo(): FolderInfo {
     }
 }
 
-export async function writeLogs() {
+export async function writeAndShowBuildLogs(isLocalInstall: boolean = false) {
     const logFilePath = path.join(os.tmpdir(), 'build-logs.txt')
-    writeFileSync(logFilePath, transformByQState.getErrorLog())
+    writeFileSync(logFilePath, transformByQState.getBuildLog())
+    const doc = await vscode.workspace.openTextDocument(logFilePath)
+    if (
+        !transformByQState.getBuildLog().includes('clean install succeeded') &&
+        transformByQState.getTransformationType() !== TransformationType.SQL_CONVERSION
+    ) {
+        // only show the log if the build failed; show it in second column for intermediate builds only
+        const options = isLocalInstall ? undefined : { viewColumn: vscode.ViewColumn.Two }
+        await vscode.window.showTextDocument(doc, options)
+    }
     return logFilePath
+}
+
+export async function createLocalBuildUploadZip(baseDir: string, exitCode: number | null, stdout: string) {
+    const manifestFilePath = path.join(baseDir, 'manifest.json')
+    const buildResultsManifest = {
+        capability: 'CLIENT_SIDE_BUILD',
+        exitCode: exitCode,
+        commandLogFileName: 'build-output.log',
+    }
+    const formattedManifest = JSON.stringify(buildResultsManifest)
+    await fs.writeFile(manifestFilePath, formattedManifest)
+
+    const buildLogsFilePath = path.join(baseDir, 'build-output.log')
+    await fs.writeFile(buildLogsFilePath, stdout)
+
+    const zip = new AdmZip()
+    zip.addLocalFile(buildLogsFilePath)
+    zip.addLocalFile(manifestFilePath)
+
+    const zipPath = `${baseDir}.zip`
+    zip.writeZip(zipPath)
+    getLogger().info(`CodeTransformation: created local build upload zip at ${zipPath}`)
+    return zipPath
+}
+
+// extract the 'sources' directory of the upload ZIP so that we can apply the diff.patch to a copy of the source code
+export async function extractOriginalProjectSources(destinationPath: string) {
+    const zip = new AdmZip(transformByQState.getPayloadFilePath())
+    const zipEntries = zip.getEntries()
+    for (const zipEntry of zipEntries) {
+        if (zipEntry.entryName.startsWith('sources')) {
+            zip.extractEntryTo(zipEntry, destinationPath, true, true)
+        }
+    }
 }
 
 export async function checkBuildSystem(projectPath: string) {
@@ -76,6 +119,17 @@ export async function parseBuildFile() {
     return undefined
 }
 
+export async function validateCustomVersionsFile(fileContents: string) {
+    const requiredKeys = ['dependencyManagement:', 'identifier:', 'targetVersion:']
+    for (const key of requiredKeys) {
+        if (!fileContents.includes(key)) {
+            getLogger().info(`CodeTransformation: .YAML file is missing required key: ${key}`)
+            return false
+        }
+    }
+    return true
+}
+
 export async function validateSQLMetadataFile(fileContents: string, message: any) {
     try {
         const sctData = await xml2js.parseStringPromise(fileContents)
@@ -119,20 +173,10 @@ export async function validateSQLMetadataFile(fileContents: string, message: any
     return true
 }
 
-export async function setMaven() {
-    let mavenWrapperExecutableName = isWin() ? 'mvnw.cmd' : 'mvnw'
-    const mavenWrapperExecutablePath = path.join(transformByQState.getProjectPath(), mavenWrapperExecutableName)
-    if (existsSync(mavenWrapperExecutablePath)) {
-        if (mavenWrapperExecutableName === 'mvnw') {
-            mavenWrapperExecutableName = './mvnw' // add the './' for non-Windows
-        } else if (mavenWrapperExecutableName === 'mvnw.cmd') {
-            mavenWrapperExecutableName = '.\\mvnw.cmd' // add the '.\' for Windows
-        }
-        transformByQState.setMavenName(mavenWrapperExecutableName)
-    } else {
-        transformByQState.setMavenName('mvn')
-    }
-    getLogger().info(`CodeTransformation: using Maven ${transformByQState.getMavenName()}`)
+export function setMaven() {
+    // for now, just use regular Maven since the Maven executables can
+    // cause permissions issues when building if user has not ran 'chmod'
+    transformByQState.setMavenName('mvn')
 }
 
 export async function openBuildLogFile() {