+++ packages/core/src/test/shared/utilities/processUtils.test.ts
@@ -10,8 +10,10 @@ import * as sinon from 'sinon'
 import { makeTemporaryToolkitFolder, tryRemoveFolder } from '../../../shared/filesystemUtilities'
 import {
     ChildProcess,
+    ChildProcessOptions,
     ChildProcessResult,
     ChildProcessTracker,
+    defaultProcessWarnThresholds,
     eof,
     ProcessStats,
 } from '../../../shared/utilities/processUtils'
@@ -376,8 +378,8 @@ async function stopAndWait(runningProcess: RunningProcess): Promise<void> {
     await runningProcess.result
 }
 
-function startSleepProcess(timeout: number = 90): RunningProcess {
-    const childProcess = new ChildProcess(getSleepCmd(), [timeout.toString()])
+function startSleepProcess(options?: ChildProcessOptions, timeout: number = 90): RunningProcess {
+    const childProcess = new ChildProcess(getSleepCmd(), [timeout.toString()], options)
     const result = childProcess.run().catch(() => assert.fail('sleep command threw an error'))
     return { childProcess, result }
 }
@@ -454,12 +456,12 @@ describe('ChildProcessTracker', function () {
         tracker.add(runningProcess.childProcess)
 
         const highCpu: ProcessStats = {
-            cpu: ChildProcessTracker.thresholds.cpu + 1,
+            cpu: defaultProcessWarnThresholds.cpu + 1,
             memory: 0,
         }
         const highMemory: ProcessStats = {
             cpu: 0,
-            memory: ChildProcessTracker.thresholds.memory + 1,
+            memory: defaultProcessWarnThresholds.memory + 1,
         }
 
         usageMock.returns(highCpu)
@@ -480,7 +482,7 @@ describe('ChildProcessTracker', function () {
         tracker.add(runningProcess.childProcess)
 
         usageMock.returns({
-            cpu: ChildProcessTracker.thresholds.cpu + 1,
+            cpu: defaultProcessWarnThresholds.cpu + 1,
             memory: 0,
         })
 
@@ -492,15 +494,63 @@ describe('ChildProcessTracker', function () {
 
     it('does not log for processes within threshold', async function () {
         const runningProcess = startSleepProcess()
+        tracker.add(runningProcess.childProcess)
 
         usageMock.returns({
-            cpu: ChildProcessTracker.thresholds.cpu - 1,
-            memory: ChildProcessTracker.thresholds.memory - 1,
+            cpu: defaultProcessWarnThresholds.cpu - 1,
+            memory: defaultProcessWarnThresholds.memory - 1,
         })
 
         await clock.tickAsync(ChildProcessTracker.pollingInterval)
 
         assert.throws(() => assertLogsContain(runningProcess.childProcess.pid().toString(), false, 'warn'))
+
+        await stopAndWait(runningProcess)
+    })
+
+    it('respects custom thresholds', async function () {
+        const largeRunningProcess = startSleepProcess({
+            warnThresholds: {
+                cpu: defaultProcessWarnThresholds.cpu + 10,
+                memory: defaultProcessWarnThresholds.memory + 10,
+            },
+        })
+        tracker.add(largeRunningProcess.childProcess)
+        const smallRunningProcess = startSleepProcess({
+            warnThresholds: {
+                cpu: defaultProcessWarnThresholds.cpu - 10,
+                memory: defaultProcessWarnThresholds.memory - 10,
+            },
+        })
+        tracker.add(smallRunningProcess.childProcess)
+
+        usageMock.returns({
+            cpu: defaultProcessWarnThresholds.cpu + 5,
+            memory: defaultProcessWarnThresholds.memory + 5,
+        })
+
+        await clock.tickAsync(ChildProcessTracker.pollingInterval)
+        assert.throws(() => assertLogsContain(largeRunningProcess.childProcess.pid().toString(), false, 'warn'))
+        assertLogsContain(smallRunningProcess.childProcess.pid().toString(), false, 'warn')
+
+        await stopAndWait(largeRunningProcess)
+        await stopAndWait(smallRunningProcess)
+    })
+
+    it('fills custom thresholds with default', async function () {
+        const runningProcess = startSleepProcess({
+            warnThresholds: {
+                cpu: defaultProcessWarnThresholds.cpu + 10,
+            },
+        })
+        tracker.add(runningProcess.childProcess)
+
+        usageMock.returns({
+            memory: defaultProcessWarnThresholds.memory + 1,
+        })
+
+        await clock.tickAsync(ChildProcessTracker.pollingInterval)
+        assertLogsContain(runningProcess.childProcess.pid().toString(), false, 'warn')
 
         await stopAndWait(runningProcess)
     })